# backend/openapi/v1.yaml
openapi: 3.0.3
info:
  title: DataAfrik API
  version: 1.0.0
  description: >
    API-first contract for DataAfrik domain services. This v1 spec is a migration
    baseline and is expanded incrementally while legacy endpoints remain active.
servers:
  - url: /api/v1
...
paths:
...
  /blog/posts:
    get:
      tags: [blog]
      summary: List blog posts
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: Posts list
  /blog/posts/{slug}:
    get:
      tags: [blog]
      summary: Fetch a post by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Post resolved
        '404':
          description: Not found
  /blog/posts/{slug}/like:
    post:
      tags: [blog]
      summary: Like a blog post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Like accepted
  /comments/post/{postId}:
    get:
      tags: [comments]
      summary: List comments for a post
      parameters:
        - in: path
          name: postId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Comment list
  /comments:
    post:
      tags: [comments]
      summary: Create comment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Comment created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...